requirements.txt
flask==2.0.1
boto3==1.26.137
sagemaker==2.151.0
python-dotenv==1.0.0
requests==2.31.0